player movement

+++ GameObject.hpp
@@ -5,13 +5,14 @@
 
 class GameObject
 {
-    private:
-        Position _position;
-    public:
-        GameObject(Position p);
-        virtual void update(int deltaTime) = 0;
-        virtual void render() = 0;
-        void setPosition(const Position& p);
-        Position getPosition();
+private:
+    Position _position;
+
+public:
+    GameObject(Position p);
+    virtual void update(float deltaTime) = 0;
+    virtual void render() = 0;
+    void setPosition(const Position &p);
+    Position getPosition();
 };
 #endif
